requirements.txt
Django==4.2.24
gunicorn==23.0.0
whitenoise==6.11.0
dj-database-url==3.0.1
psycopg2-binary==2.9.11
python-decouple==3.8
stripe==14.0.1
django-import-export==4.3.10
tablib==3.8.0
pillow==10.4.0
reportlab==4.4.5
xhtml2pdf==0.2.17
svglib==1.6.0
lxml==6.0.2
html5lib==1.1
tinycss2==1.4.0
cssselect2==0.8.0


